Split decorators into multiple files, add typedocs
