+++ src/features/milestones/milestone.tsx
@@ -1,5 +1,5 @@
 import Select from "components/fields/Select.vue";
-import { Decorator } from "features/decorators";
+import { Decorator } from "features/decorators/common";
 import type {
     CoercableComponent,
     GenericComponent,
